pyproject.toml
[project]
name = "game-predictor"
version = "0.1.0"
description = "Gemma 4 26B agentic web scraper & DuckDB pipeline for video game classification model training"
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
dependencies = [
    "duckdb>=1.2.0",
    "playwright>=1.49.0",
    "langchain-openai>=0.3.0",
    "langgraph>=0.4.0",
    "pydantic>=2.10.0",
    "aiohttp>=3.11.0",
    "aiofiles>=24.1.0",
    "httpx>=0.27.0",
    "Pillow>=11.0.0",
    "rich>=13.9.0",
    "tenacity>=9.0.0",
    "imagehash>=4.3.0",
]

[project.optional-dependencies]
notebooks = [
    "pandas>=2.2.0",
    "matplotlib>=3.9.0",
    "seaborn>=0.13.0",
    "numpy>=2.1.0",
    "ipykernel>=6.29.0",
]
nn = [
    "torch>=2.2.0",
    "scikit-learn>=1.4.0",
    "numpy>=2.1.0",
]
cnn = [
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "scikit-learn>=1.4.0",
    "numpy>=2.1.0",
    "opencv-python>=4.9.0",
]
transformer = [
    "sentence-transformers>=3.0.0",
    "torch>=2.2.0",
    "scikit-learn>=1.4.0",
    "numpy>=2.1.0",
]

[project.scripts]
game-predictor-test = "game_predictor.cli.test_run:cli_main"
game-predictor-run = "game_predictor.cli.run_production:cli_main"
game-predictor-preflight = "game_predictor.cli.pre_flight:main"
game-predictor-sleeper = "game_predictor.cli.sleeper:main"
game-nn-train = "game_nn.pipeline:run"
game-cnn-train = "game_cnn.pipeline:run"
game-transformer-train = "game_transformer.pipeline:run"
game-ensemble = "ensemble.run:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/game_predictor", "src/game_nn", "src/game_cnn", "src/game_transformer", "src/ensemble"]
